requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.2
boto3==1.12.39
botocore==1.15.39
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-import-export==2.0.2
django-material-admin==1.7.0
django-progressbarupload==0.1.7
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
jdcal==1.4.1
jmespath==0.9.5
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.1
pipenv==2018.11.26
psycopg2==2.8.5
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
stripe==2.48.0
tablib==1.1.0
typed-ast==1.4.1
Unipath==1.1
urllib3==1.25.8
virtualenv==20.0.18
virtualenv-clone==0.5.4
whitenoise==5.0.1
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
